refactor: move is_accelerator_type to utils/helpers
Address PR #665 review comments from kylesayrs:
- Relocate is_accelerator_type to compressed_tensors.utils.helpers
  since it's a general device utility, not specific to offload/convert
- Remove unused _accel_type() helper
- Move inline import to top-level in disk.py
- Update all import sites (base.py, test_xpu_routing.py)

Refs: #665
