+++ tests/test_offload/test_xpu_routing.py
@@ -17,7 +17,8 @@ import pytest
 import torch
 from compressed_tensors.offload.cache.base import OffloadCache
 from compressed_tensors.offload.cache.device import DeviceCache
-from compressed_tensors.offload.convert.helpers import is_accelerator_type, norm_device
+from compressed_tensors.utils.helpers import is_accelerator_type
+from compressed_tensors.offload.convert.helpers import norm_device
 
 
 @pytest.fixture
